Made many changes

+++ task01/src/com/example/PlayStoreApp/PlaystoreAppData.java
@@ -3,10 +3,10 @@ package com.example.PlayStoreApp;
 public class PlaystoreAppData {
 
   private String appName;
-  private double rating;
+  private Double rating;
   private String category;
 
-  public PlaystoreAppData(String appName, double rating, String category) {
+  public PlaystoreAppData(String appName, Double rating, String category) {
     this.appName = appName;
     this.rating = rating;
     this.category = category;
@@ -20,9 +20,9 @@ public class PlaystoreAppData {
 
   public void setAppName(String appName) { this.appName = appName; }
 
-  public double getRating() { return rating; }
+  public Double getRating() { return rating; }
 
-  public void setRating(Float rating) { this.rating = rating; }
+  public void setRating(Double rating) { this.rating = rating; }
 
   public String getCategory() { return category; }
